feat(site): consolidate 3 pitch pages into one /teams (#656)
One clean canonical page at /teams, rebuilt from the strongest (post-roast)
content and folding the skill-mesh angle in as a single section:
hero -> zero-data + architecture diagram -> rollup mock (+ live demo CTA)
-> sample payload -> how it works -> spread-what-works / skill mesh
-> k-anonymity -> security FAQ -> founding cohort (4 of 5 open) -> pricing -> CTA.

Retire /design-partners and /registry: delete the routes, 301 both to /teams
(before the SPA fallback in _redirects), repoint internal links, drop PitchSwitch.

Verified: build (prerenders /teams, no dangling links) + typecheck pass.
